requirements.txt
aiogram==3.0.0b7
async-timeout==4.0.2
asynccpu==1.2.2
asyncffmpeg==1.2.0
decorator==4.4.2
ffmpeg-python==0.2.0
moviepy==1.0.3
numpy==1.23.1
opencv-python==4.6.0.66
piexif==1.1.3
Pillow==9.2.0
proglog==0.1.10
psutil==5.9.1